refactor: consolidate assertNever and waitForRecordedEvent into shared utilities
Move assertNever from packages/cli/src/chat/tui/assertNever.ts into
src/utils/core/typeGuards.ts (already the home for isObject, unique, etc.)
and export it through the @utils/core barrel. Update all six CLI callers
and the one Desktop caller to import from @utils/core; delete the now-empty
local file. The Desktop variant had a narrower signature (no message param)
and inline error text — both consolidated into the canonical two-arg form.

Extract the identical waitForRecordedEvent helper defined in
WolframToolApproval.vitest.ts and HumanPromptProgressEvents.vitest.ts into
src/test-kernel/support/asyncTestUtils.ts and import from there.

Net: -42 lines, one fewer file, two fewer copy-paste duplicates.
