+++ src/test-kernel/agent/followUp/HumanPromptProgressEvents.vitest.ts
@@ -26,18 +26,7 @@ import {
   setToolEditApprovalHandler,
 } from '@tools/approval/toolEditApproval';
 import { createRecordingHost } from '../progressTestUtils';
-
-async function waitForRecordedEvent<TEvent extends string>(
-  events: Array<{ event: TEvent; payload: unknown }>,
-  eventName: TEvent,
-): Promise<{ event: TEvent; payload: any }> {
-  for (let attempt = 0; attempt < 10; attempt++) {
-    const event = events.find((entry) => entry.event === eventName);
-    if (event) return event;
-    await new Promise((resolve) => setImmediate(resolve));
-  }
-  throw new Error(`Timed out waiting for ${eventName}`);
-}
+import { waitForRecordedEvent } from '../../support/asyncTestUtils';
 
 describe('human prompt progress events', () => {
   afterEach(() => {
