+++ src/test-kernel/tools/WolframToolApproval.vitest.ts
@@ -16,18 +16,7 @@ import * as wolframScriptUtils from '@tools/wolfram/wolframScriptUtils';
 
 // Local imports - test helpers
 import { createRecordingHost } from '../agent/progressTestUtils';
-
-async function waitForRecordedEvent<TEvent extends string>(
-  events: Array<{ event: TEvent; payload: unknown }>,
-  eventName: TEvent,
-): Promise<{ event: TEvent; payload: any }> {
-  for (let attempt = 0; attempt < 10; attempt++) {
-    const event = events.find((entry) => entry.event === eventName);
-    if (event) return event;
-    await new Promise((resolve) => setImmediate(resolve));
-  }
-  throw new Error(`Timed out waiting for ${eventName}`);
-}
+import { waitForRecordedEvent } from '../support/asyncTestUtils';
 
 describe('WolframTool approval', () => {
   afterEach(() => {
